requirements.txt
pyssdeep==1.0.0
nltk==3.8.1
flyhash==1.1.1
gensim==4.3.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
datasketch==1.5.9
fuzzywuzzy==0.18.0
idna==3.4
joblib==1.3.1
Levenshtein==0.21.1
python-Levenshtein==0.21.1
rapidfuzz==3.1.1
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
threadpoolctl==3.2.0
tqdm==4.65.0
urllib3==2.0.3
et-xmlfile
jieba
simhash
numpy
openpyxl
python-mmdt
ssdeep-windows-32-64